[TASK] Fixes for WallsService to fix issues

+++ Classes/Service/WallsService.php
@@ -22,7 +22,6 @@ use TYPO3\CMS\Core\Http\ServerRequestFactory;
 use TYPO3\CMS\Core\Registry;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
-use TYPO3\CMS\Core\Utility\StringUtility;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
@@ -51,8 +50,8 @@ class WallsService
         'post_link',
     ];
 
+    protected string $targetDirectory = 'typo3temp/assets/walls_io_proxy';
     protected Registry $registry;
-
     protected WallsIoClient $client;
 
     public function __construct(Registry $registry, WallsIoClient $client)
@@ -228,7 +227,7 @@ class WallsService
 
         if (
             array_key_exists('external_image', $post)
-            && StringUtility::beginsWith((string)$post['external_image'], 'http')
+            && str_starts_with((string)$post['external_image'], 'http')
         ) {
             $post['external_image'] = $this->cacheExternalResources(
                 $post['external_image'],
@@ -238,7 +237,7 @@ class WallsService
 
         if (
             array_key_exists('post_image', $post)
-            && StringUtility::beginsWith((string)$post['post_image'], 'http')
+            && str_starts_with((string)$post['post_image'], 'http')
         ) {
             $post['post_image'] = $this->cacheExternalResources(
                 $post['post_image'],
@@ -257,7 +256,7 @@ class WallsService
                 && is_array($matches['src'])
             ) {
                 foreach ($matches['src'] as $uri) {
-                    if (StringUtility::beginsWith($uri, 'http')) {
+                    if (str_starts_with($uri, 'http')) {
                         $post['comment'] = str_replace(
                             $matches['src'],
                             $this->cacheExternalResources($uri, $pluginConfiguration->getRecordUid()),
@@ -275,6 +274,7 @@ class WallsService
     protected function cacheExternalResources(string $resource, int $contentRecordUid): string
     {
         $pathParts = GeneralUtility::split_fileref(parse_url($resource, PHP_URL_PATH));
+
         $filePath = sprintf(
             '%s%s.%s',
             $this->getTargetDirectory($contentRecordUid),
@@ -282,7 +282,7 @@ class WallsService
             $pathParts['fileext']
         );
 
-        if (!file_exists($filePath)) {
+        if (!file_exists($filePath) && GeneralUtility::getUrl($resource)) {
             GeneralUtility::writeFile($filePath, GeneralUtility::getUrl($resource));
         }
